+++ lib/services/portal_server/commands/portal_merge.ex
@@ -4,9 +4,11 @@ defmodule Sorcery.PortalServer.Commands.PortalMerge do
 
   def entry(%{portal: portal}, %Sorcery.PortalServer.InnerState{} = inner_state) do
     %{portal_name: name} = portal
+    pending_portals = List.delete(inner_state.pending_portals, name)
 
     inner_state
     |> put_in_p([:portals, name], portal)
+    |> put_in_p([:pending_portals], pending_portals)
   end